+++ app/src/main/java/com/bionic/td_android/Data/Provider/job/JobModel.java
@@ -1,8 +1,11 @@
 package com.bionic.td_android.Data.Provider.job;
 
-import android.support.annotation.Nullable;
-
 import com.bionic.td_android.Data.Provider.base.BaseModel;
+
+import java.util.Date;
+
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 
 /**
  * Job entity
@@ -21,5 +24,5 @@ public interface JobModel extends BaseModel {
      * Can be {@code null}.
      */
     @Nullable
-    Integer getJobid();
+    Long getJobid();
 }
